Extract app page and search page navbar
